+++ src/main/index.ts
@@ -1,7 +1,7 @@
 console.time('init')
 
-import { ensureDirSync, watchFile, readFile } from 'fs-extra'
-import { app as rawApp, dialog, protocol } from 'electron'
+import { ensureDirSync, watchFile } from 'fs-extra'
+import { app as rawApp, dialog } from 'electron'
 import rc from './rc'
 import { VERSION, GIT_REF, BUILD_TIMESTAMP } from '../shared/build-info'
 import type { EventEmitter } from 'events'
@@ -26,7 +26,6 @@ import {
   getLogsPath,
   getAccountsPath,
   getCustomThemesPath,
-  htmlDistDir,
 } from './application-constants'
 ensureDirSync(getConfigPath())
 ensureDirSync(getLogsPath())
@@ -69,8 +68,6 @@ import { AppState, DeltaChatAccount } from '../shared/shared-types'
 import { ExtendedAppMainProcess } from './types'
 import { updateTrayIcon, hideDeltaChat, showDeltaChat } from './tray'
 import { acceptThemeCLI } from './themes'
-import { join, normalize, sep, extname, basename } from 'path'
-import { lookup } from 'mime-types'
 
 app.ipcReady = false
 app.isQuitting = false
@@ -198,92 +195,4 @@ app.on('web-contents-created', (_e, contents) => {
   })
 })
 
-protocol.registerSchemesAsPrivileged([
-  { scheme: 'dc', privileges: { standard: true } },
-  { scheme: 'dc-blob', privileges: { stream: true } },
-  { scheme: 'misc', privileges: { stream: true } },
-])
-
-// folders the renderer need to load resources from
-const ALLOWED_RESOURCE_FOLDERS = ['images', 'node_modules']
-// folders the renderer wants to load source files from (when using the devtools)
-const ALLOWED_SOURCE_FOLDERS = ['src', 'scss', 'node_modules']
-const ALLOWED_FOLDERS = [...ALLOWED_RESOURCE_FOLDERS, ...ALLOWED_SOURCE_FOLDERS]
-const BASE_DIR = join(htmlDistDir(), '../')
-const HTML_DIST_DIR = htmlDistDir()
-
-const ACCOUNTS_DIR = getAccountsPath()
-
-app.once('ready', () => {
-  protocol.registerBufferProtocol('misc', async (req, cb) => {
-    if (req.url.startsWith('misc://background/')) {
-      const filename = basename(req.url.replace('misc://background/', ''))
-      readFile(join(getConfigPath(), 'background/', filename), (e, b) => {
-        if (e) {
-          log.warn('error while fetching background image', filename, e)
-          cb({ statusCode: 404 })
-        } else {
-          cb(b)
-        }
-      })
-    } else {
-      cb({ statusCode: 400 })
-    }
-  })
-  protocol.registerBufferProtocol('dc-blob', (req, cb) => {
-    // check for path escape attempts
-    const file = normalize(req.url.replace('dc-blob://', ''))
-    if (file.indexOf('..') !== -1) {
-      log.warn('path escape prevented', req.url, file)
-      return cb({ statusCode: 400 })
-    }
-
-    // Fetch Blobfile - make sure its really in a blob dir
-    if (!file.split(sep).includes('db.sqlite-blobs')) {
-      log.warn(
-        'error while fetching blob file - id not inside the blobs directory',
-        file
-      )
-      cb({ statusCode: 400 })
-    } else {
-      readFile(join(ACCOUNTS_DIR, file.replace('p40', 'P40')), (e, b) => {
-        if (e) {
-          log.warn('error while fetching blob file', file, e)
-          cb({ statusCode: 404 })
-        } else {
-          cb(b)
-        }
-      })
-    }
-  })
-  protocol.registerBufferProtocol('dc', (req, cb) => {
-    // check for path escape attempts
-    let file = normalize(req.url.replace('dc://deltachat/', ''))
-    const hashtagIndex = file.indexOf('#')
-    if (hashtagIndex !== -1) {
-      file = file.slice(0, hashtagIndex)
-    }
-    if (file.indexOf('..') !== -1) {
-      log.warn('path escape prevented', req.url, file)
-      return cb({ statusCode: 400 })
-    }
-
-    const otherFolder = ALLOWED_FOLDERS.find(folder =>
-      file.startsWith(folder + '/')
-    )
-    const prefix = otherFolder ? BASE_DIR : HTML_DIST_DIR
-
-    // Fetch resource or source
-    readFile(join(prefix, file.replace(/:$/, '')), (e, b) => {
-      if (e) {
-        log.warn('error while fetching resource', file, e)
-        cb({ statusCode: 404 })
-      } else {
-        cb({
-          mimeType: lookup(extname(file.replace(/:$/, ''))) || undefined,
-          data: b,
-        })
-      }
-    })
-  })
-})
+import('./internal-app-schemes')
